Added support for __BenchmarkCrossDomainData
There was not a way to share data between tests. It should not be done
but if initialization involves some very expensive operation (which
can't be performed outside, before calling benchmark engine) then such
mechanism was required. Solution is to share a MarshalByRefObject
between main AppDomain with each newly created one.